myanimelist/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
PyYAML==6.0.1
uvicorn==0.34.0
google-cloud-storage==3.0.0
fastapi==0.115.8
aiohttp==3.11.12
tenacity==9.0.0